chore: #102 p2p- move utils to separate file
