+++ src/node/p2p/mod.rs
@@ -1,15 +1,11 @@
 pub mod error;
 mod getters;
 mod types;
+mod utils;
 // external
-use std::{
-    io::{self, Read},
-    net::{IpAddr, SocketAddr, TcpListener, TcpStream},
-    time::Duration,
-};
+use std::net::{IpAddr, SocketAddr, TcpListener};
 // local
 pub use error::P2PError;
-use types::Result;
 
 #[derive(Debug)]
 pub struct P2P {
@@ -19,35 +15,8 @@ pub struct P2P {
     listener: TcpListener,
 }
 
-fn handle_stream(mut stream: TcpStream) -> std::result::Result<(), P2PError> {
-    let mut buf = [0];
-    stream.set_read_timeout(Some(Duration::from_millis(100)))?;
-
-    loop {
-        let _ = match stream.read(&mut buf) {
-            Err(e) => match e.kind() {
-                io::ErrorKind::WouldBlock => {
-                    println!("Would have blocked");
-                    break;
-                }
-                _ => panic!("Got an error"),
-            },
-            Ok(m) => {
-                println!("Received {m:?}, {buf:?}");
-                if m == 0 {
-                    break;
-                };
-                m
-            }
-        };
-    }
-
-    Ok(())
-}
-
 impl P2P {
-    /// ## Initialize a new peer-to-peer connection
-    ///
+    /// ### Initialize a new peer-to-peer connection
     /// Get environment information from a config file.
     pub fn new(host: IpAddr, port: u16) -> Self {
         let socket_addr = SocketAddr::new(host, port);
@@ -59,27 +28,5 @@ impl P2P {
             socket_addr,
             listener,
         }
-    }
-    pub fn start_connection(&self) -> Result<()> {
-        // self.listener.set_nonblocking(true)?;
-
-        for stream_res in self.listener.incoming() {
-            match stream_res {
-                Ok(s) => handle_stream(s)?,
-                Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => {
-                    break;
-                }
-                Err(e) => panic!("Unknown error: {}", e),
-            }
-        }
-
-        Ok(())
-    }
-    /// ### Start the p2p connection.
-    pub fn start_p2p(&mut self) -> Result<()> {
-        // start the connection
-        self.start_connection()?;
-
-        Ok(())
     }
 }
